Wrote the code for both of the functions, encrypt, and decrypt.
Encrypt will take in the letters of the message, and then will see if the letter is in a sipher. It will then replace the letter with the letter in the sipher.

Decrypt will take in the letters of the message, and then will see if the letter is in a sipher. It will then replace the letter with the letter in the sipher.